kauriac001/terraform-mistralaisuite-gcp-v0.1.3/modules/calculator/pyproject.toml
[project]
name = "calculator"
version = "0.1.0"
description = "Cloud Resource Calculator"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "azure-identity>=1.23.1",
    "azure-mgmt-compute>=35.0.0",
    "boto3>=1.40.1",
    "fire>=0.7.0",
    "google-auth>=2.40.3",
    "google-cloud-compute>=1.33.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "ruff>=0.12.7",
]
